requirements.txt
openai>=1.30.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
